feat: saved pog working config
This has a funcional keymap with only one layer. It doesn't work because
some config makes the keymap have many unused transparent keys and some
allocation issues arise from it's poor memory usage. BUT, it's the first
config that functioned in my keyboard setup so I will save it for
reference.